requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
boto3==1.16.51
botocore==1.19.51
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.4
django-allauth==0.44.0
django-braces==1.14.0
django-embed-video==1.4.0
django-heroku==0.3.1
django-storages==1.11.1
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
isort==5.6.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.0.1
praw==7.1.0
prawcore==1.5.0
psaw==0.0.12
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
update-checker==0.18.0
uritools==3.0.0
urlextract==1.1.0
urllib3==1.25.11
websocket-client==0.57.0
whitenoise==5.2.0
wikipedia==1.4.0
wrapt==1.12.1
